refactor: move extractQuotedKey to dedicated module
Moves extractQuotedKey and tryMatchAsClosingQuote from registerLanguage.ts
into a dedicated extractQuotedKey.ts module. This decouples the pure string
helper from the Monaco registration wiring, making tests less brittle and
enabling easier reuse.
